chore(ci): make the scoped lint gate blocking; narrow the provenance fallback
The lint step was `continue-on-error` while the paths it covered still carried 12
findings, which made it decorative — the same shape of defect as a guard that
cannot fire. Findings fixed, so the gate now blocks, and ruff is pinned (0.16.2)
because an unpinned blocking linter turns CI red on a release with no code
change, which is how a gate gets disabled a second time.

- `src/metrics_io.py`: `git_commit()` no longer swallows bare `Exception` — only
  `OSError` (no git on PATH) and `SubprocessError` (not a repo / no commits), so
  a real bug in it surfaces instead of being relabelled "unknown". The append in
  `write_run` stays deliberately unguarded: prereg §8 makes it the single sink,
  so a failed write must propagate.
- `src/data/{tiling,multiyear}.py`: the block_px arithmetic was open-coded in
  four places; route all of them through the existing `block_px_from_cfg`.
- `tests/`: ruff cleanups only, each value-identical (`itertools.pairwise`, dict
  literals, redundant `int()` around `round()`, unused import).

`pytest tests/ -q` -> 112 passed. `ruff check` on the scoped paths -> clean.

+++ tests/synthetic.py
@@ -129,8 +129,11 @@ def old_corner_rule_splits(tile_xy, block_px: int = BLOCK_PX, ratios: dict = RAT
     rng = random.Random(seed)
     rng.shuffle(blocks)
     n = len(blocks)
-    n_train = max(int(round(ratios["train"] * n)), 1 if n else 0)
-    n_val = int(round(ratios["val"] * n))
+    # round() already returns a Python int here (ratios are YAML floats, n is a
+    # len()), so no int() cast — value-identical, and the CI negative-control step
+    # plus test_old_corner_rule_would_leak both re-prove this rule still leaks.
+    n_train = max(round(ratios["train"] * n), 1 if n else 0)
+    n_val = round(ratios["val"] * n)
     assign = {}
     for i, b in enumerate(blocks):
         assign[b] = "train" if i < n_train else ("val" if i < n_train + n_val else "test")
